Add files via upload
In the above project, I tested the Le-net architecture on the MNIST dataset. 
Well, I've got and testing accuracy of 99 percent. I read a bit about the original Le-net architecture of how they did on NIST handwritten dataset.
I chose the Adam optimizer with default settings , but with a learning rate of 0.00025. Still, we can see a lot of oscillations on the loss curve , which suggests that
we got to tune the beta1 and beta2, to a bit low value.
We can also here see the output of each channels in the convolutional network.  How the images are being processed in the convolutional network, is pretty amazing. :)